+++ estimators/tangentBC.cpp
@@ -39,6 +39,7 @@
 
 #include "DGtal/base/Common.h"
 #include "DGtal/helpers/StdDefs.h"
+#include "DGtal/io/readers/PointListReader.h"
 
 //Grid curve
 #include "DGtal/geometry/curves/FreemanChain.h"
@@ -65,7 +66,7 @@ namespace po = boost::program_options;
  @b Allowed @b options @b are : 
  @code
   -h [ --help ]              display this message
-  -i [ --input ] arg         input FreemanChain file name
+  -i [ --input ] arg         input file name: FreemanChain (.fc) or a sequence of discrete points (.sdp).
   -s [ --GridStep ] arg (=1) Grid step
  @endcode
 
@@ -105,7 +106,7 @@ int main( int argc, char** argv )
   po::options_description general_opt("Allowed options are: ");
   general_opt.add_options()
     ("help,h", "display this message")
-    ("input,i", po::value<std::string>(), "input FreemanChain file name")
+    ("input,i", po::value<std::string>(), "input file name: FreemanChain (.fc) or a sequence of discrete points (.sdp).")
     ("GridStep,step", po::value<double>()->default_value(1.0), "Grid step");
   
   
@@ -135,7 +136,8 @@ int main( int argc, char** argv )
  
   if(vm.count("input")){
     std::string fileName = vm["input"].as<std::string>();
-
+    std::string extension =  fileName.substr( fileName.find_last_of(".") + 1 );
+    bool isSDP = extension == "sdp";
     typedef Z2i::Space Space; 
     typedef Space::Point Point; 
     typedef PointVector<2, double> RealPoint; 
@@ -144,18 +146,30 @@ int main( int argc, char** argv )
     typedef std::vector< Point > Storage;
     typedef Storage::const_iterator ConstIteratorOnPoints; 
 
-    std::vector< FreemanChain > vectFcs =  
-      PointListReader< Point >:: getFreemanChainsFromFile<Integer> (fileName); 
-    
-    for(unsigned int i=0; i<vectFcs.size(); i++){
-
-      bool isClosed = vectFcs.at(i).isClosed(); 
-      std::cout << "# grid curve " << i << "/" << vectFcs.size() << " "
-                << ( (isClosed)?"closed":"open" ) << std::endl;
-
+    std::vector< FreemanChain > vectFcs;
+    if(!isSDP)
+      {
+        vectFcs =   
+          PointListReader< Point >:: getFreemanChainsFromFile<Integer> (fileName); 
+      }
+    for(unsigned int i=0; i<vectFcs.size() || (i==0 && isSDP); i++){
       Storage vectPts; 
-      FreemanChain::getContourPoints( vectFcs.at(i), vectPts ); 
-
+      bool isClosed;
+      if(!isSDP)
+        {
+          isClosed = vectFcs.at(i).isClosed(); 
+          std::cout << "# grid curve " << i << "/" << vectFcs.size() << " "
+                    << ( (isClosed)?"closed":"open" ) << std::endl;
+          FreemanChain::getContourPoints( vectFcs.at(i), vectPts ); 
+        }
+      else
+        {
+          vectPts = PointListReader<Z2i::Point>::getPointsFromFile(fileName);
+          Z2i::Point pf =vectPts[0];
+          Z2i::Point pl =vectPts[vectPts.size()-1];
+          isClosed = (pf[0]-pl[0])+(pf[1]-pl[1]) <= 1;
+        }
+      
       // Binomial
       std::cout << "# Curvature estimation from binomial convolution" << std::endl;
       typedef BinomialConvolver<ConstIteratorOnPoints, double> MyBinomialConvolver;
@@ -174,7 +188,7 @@ int main( int argc, char** argv )
              tangents.begin() ); 
 
       // Output
-      std::cout << "# id tangent.x tangent.y angle(atan2(y,x))" << std::endl;  
+      std::cout << "# id tangent.x tangent.y angle(atan2(y,x)) x y" << std::endl;  
       unsigned int j = 0;
       for ( ConstIteratorOnPoints 
         it = vectPts.begin(), it_end = vectPts.end();
@@ -184,7 +198,7 @@ int main( int argc, char** argv )
     double y = tangents[ j ][ 1 ];
     std::cout << j << std::setprecision( 15 )
               << " " << x << " " << y 
-              << " " << atan2( y, x )
+              << " " << atan2( y, x ) << " " << (*it)[0] << " " << (*it)[1]
               << std::endl;
   }