+++ ChangeLog.md
@@ -28,7 +28,8 @@
 - *estimators*:
     - 2dlocalEstimators: add an option to export the generated contour.
      (Bertrand Kerautret, [#285](https://github.com/DGtal-team/DGtalTools/pull/285))
-    
+    - tangentBC: add an option to read sdp points as input.
+     (Bertrand Kerautret, [#285](https://github.com/DGtal-team/DGtalTools/pull/288))
     
     
 # DGtalTools 0.9.2